requirements.txt
Django==1.6.1
South==0.8.4
django-bootstrap3==3.1.0
django-braces==1.4.0
django-extensions==1.3.3
django-webtest==1.7.6
factory-boy==2.3.1
mock==1.0.1
